requirements.txt
pandas
fastapi
uvicorn
sentence-transformer
mysql-connector-python
streamlit
pydantic
python-dotenv
